+++ config/routes.php
@@ -2,13 +2,14 @@
 declare(strict_types=1);
 
 use BrowscapSite\Handler\DownloadHandler;
+use BrowscapSite\Handler\StatsHandler;
 use BrowscapSite\Handler\StreamHandler;
 use BrowscapSite\Handler\UserAgentLookupHandler;
-use Psr\Container\ContainerInterface;
 use Slim\App;
 
-return function (App $app, ContainerInterface $container): void {
+return function (App $app): void {
     $app->get('/', DownloadHandler::class);
     $app->any('/ua-lookup', UserAgentLookupHandler::class);
     $app->any('/stream', StreamHandler::class);
+    $app->any('/statistics', StatsHandler::class);
 };
