+++ src/BrowscapSiteWeb.php
@@ -6,10 +6,6 @@ class BrowscapSiteWeb
 {
     public function defineServices()
     {
-        $this['stats.controller'] =     function () {
-            return new Controller\StatsController($this, $this['pdo']);
-        };
-
         $this['version.controller'] =     function () {
             return new Controller\VersionController($this);
         };
@@ -25,7 +21,6 @@ class BrowscapSiteWeb
 
     public function defineControllers()
     {
-        $this->get('/statistics', 'stats.controller:indexAction');
         $this->get('/version', 'version.controller:indexAction');
         $this->get('/version-number', 'version.number.controller:indexAction');
         $this->get('/version.xml', 'version.xml.controller:indexAction');
